[Fix] WeaponDamage Int에서 flaot로 수정

+++ Assets/Scripts/Data/Weapon.cs
@@ -8,7 +8,7 @@ public class Weapon : Item
 {
     [Header("Weapon")]
     public MusicWeaponType Type;
-    public int[] AttackDamage;
+    public float[] AttackDamage;
     public float AttackSpeed;
     public AnimatorOverrideController WeaponOverride = null;
     public float AttackRange;
@@ -17,8 +17,7 @@ public class Weapon : Item
     public int WeaponMaxUpgrade;
     public GameObject[] WeaponParticle;
     public Projectile Projectile = null;
-    public Projectile[] Projectiles = null;
-    
+    public Projectile[] Projectiles = null;  
 }
 
 public enum MusicWeaponType
